@@ -4,23 +4,23 @@
   "private": true,
   "dependencies": {
     "@mrblenny/react-flow-chart": "^0.0.14",
-    "axios": "^0.21.1",
-    "bootstrap": "^4.6.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.5.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-spinners": "^0.10.6",
-    "styled-components": "^5.2.3",
-    "web-vitals": "^1.0.1"
+    "axios": "^1.7.7",
+    "bootstrap": "^5.3.3",
+    "react": "^18.3.1",
+    "react-bootstrap": "^2.10.5",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.27.0",
+    "react-scripts": "5.0.1",
+    "react-spinners": "^0.14.1",
+    "styled-components": "^6.1.13",
+    "web-vitals": "^4.2.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "cross-env CI=true react-scripts test --testTimeout 30000 --runInBand --testMatch \"**/*.test.js\"",
+    "test": "cross-env CI=true react-scripts test --testTimeout 30000 --runInBand --testMatch \"**/*.test.js\" --transformIgnorePatterns \"node_modules/(?!axios)/\"",
     "eject": "react-scripts eject",
-    "test:pact": "cross-env CI=true react-scripts test --testTimeout 30000 --runInBand --setupFiles ./pact/setup.js --setupFilesAfterEnv ./pact/jest-wrapper.js --testMatch \"**/*.test.pact.js\" --passWithNoTests",
+    "test:pact": "cross-env CI=true react-scripts test --testTimeout 30000 --runInBand --testMatch \"**/*.test.pact.js\" --passWithNoTests --transformIgnorePatterns \"node_modules/(?!axios)/\"",
     "publish:pact": "pact-broker publish ./pacts --consumer-app-version=1.0.0 --broker-base-url=$PACT_BROKER_URL --broker-token=$PACT_BROKER_TOKEN"
   },
   "eslintConfig": {
@@ -42,12 +42,15 @@
     ]
   },
   "devDependencies": {
-    "@pact-foundation/pact": "^9.15.5",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@emotion/react": "^11.13.3",
+    "@pact-foundation/pact": "^13.1.4",
     "@pact-foundation/pact-node": "^10.12.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/react-hooks": "^5.1.2",
-    "@testing-library/user-event": "^12.1.10",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/user-event": "^14.5.2",
     "cross-env": "^7.0.3"
   }
 }
